requirements.txt
aiohttp==3.9.1
feedparser==6.0.10
python-telegram-bot==20.7
requests==2.31.0
python-dotenv==1.0.0
openai==1.30.0